% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{acoes}
\alias{acoes}
\title{Acoes}
\format{
Uma lista que descreve as colunas:
\describe{
\item{Codigo}{Codigo sob o qual a acao e negociada}
\item{Empresa}{Nome da empresa}

}
}
\source{
\url{https://acoes.info/acoes/}
}
\usage{
acoes
}
\description{
Essa base possui todas as acoes listadas na B3
}
\keyword{datasets}
