% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analise_geral.R
\name{analise_geral}
\alias{analise_geral}
\title{Analise geral dos seus ativos}
\usage{
analise_geral(arquivo_cei = arquivo_cei, classe = acoes)
}
\arguments{
\item{arquivo_cei}{tabela xlsx contendo o extrato de negociação dos ativos de renda variavel gerada pelo site da B3}

\item{classe}{qual classe de ativo deve ser analisado (acoes, etfs ou fiis), sem aspas}
}
\value{
uma tibble
}
\description{
Essa funcao calcula o preço medio, o custo de aquisicao e a quantidade de cada ativo que a pessoa possui em carteria
}
