% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{arquivo_cei}
\alias{arquivo_cei}
\title{Arquivo CEI}
\format{
Uma lista que descreve as colunas:
\describe{
\item{codigo_de_negociacao}{Codigo sob o qual o etf e negociado}
\item{tipo_de_movimentacao}{Se a movimentacao foi de compra ou venda}
\item{data_do_negocio}{Data em que foi realizada a movimentacao}
\item{quantidade}{Numero de ativos}
\item{valor}{Valor total da movimentacao}
\item{preco}{Preco individual do ativo}

}
}
\usage{
arquivo_cei
}
\description{
Essa base serve de teste para executar as funcoes presentes neste pacote
}
\keyword{datasets}
