% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discriminacao.R
\name{discriminacao_acoes}
\alias{discriminacao_acoes}
\title{Gerando a discriminacao das acoes}
\usage{
discriminacao_acoes(cei_acoes)
}
\arguments{
\item{cei_acoes}{arquivo de saida do script analise_geral filtrado por acoes}
}
\value{
um arquivo de texto
}
\description{
Gerando a discriminacao das acoes
}
