% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discriminacao.R
\name{discriminacao_etfs}
\alias{discriminacao_etfs}
\title{Gerando a discriminacao dos etfs}
\usage{
discriminacao_etfs(b3_etfs)
}
\arguments{
\item{b3_etfs}{arquivo de saida do script analise_geral filtrado por etfs}
}
\value{
um arquivo de texto
}
\description{
Gerando a discriminacao dos etfs
}
