% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discriminacao.R
\name{discriminacao_fiis}
\alias{discriminacao_fiis}
\title{Gerando a discriminacao dos fiis}
\usage{
discriminacao_fiis(cei_fiis)
}
\arguments{
\item{cei_fiis}{arquivo de saida do script analise_geral filtrado por fiis}
}
\value{
um arquivo de texto
}
\description{
Gerando a discriminacao dos fiis
}
