% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{etfs}
\alias{etfs}
\title{ETFs}
\format{
Uma lista que descreve as colunas:
\describe{
\item{codigo}{Codigo sob o qual o etf e negociado}
\item{nome}{Nome do ETF}

}
}
\source{
\url{https://br.investing.com/etfs/brazil-etfs}
}
\usage{
etfs
}
\description{
Essa base possui todos os ETFs listados na B3
}
\keyword{datasets}
