% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{fiis}
\alias{fiis}
\title{FIIs}
\format{
Uma lista que descreve as colunas:
\describe{
\item{codigo}{Codigo sob o qual o fii e negociado}
\item{nome}{Nome do fundo de investimento imobiliario}
\item{administrador}{Nome do administrador}

}
}
\source{
\url{https://www.clubefii.com.br/fundo_imobiliario_lista}
}
\usage{
fiis
}
\description{
Essa base possui todos os FIIs listados na B3
}
\keyword{datasets}
