% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/situacao.R
\name{situacao}
\alias{situacao}
\title{Situação anual dos seus ativos}
\usage{
situacao(arquivo_b3 = arquivo_b3)
}
\arguments{
\item{arquivo_b3}{tabela xlsx contendo o extrato de negociacao dos ativos de renda variavel gerada pelo site da B}
}
\value{
uma tibble
}
\description{
Situação anual dos seus ativos
}
